FIX: Format change on gov.pl
WARNING: This is a first commit from the series with fixes.

On November 24, 2020 data format has been changed on gov.pl site, and
the script stopped working. With this commit brings back the possibility to
gather data, but display and plot function are still not operational.

Also with this commit missing data for November 24, 2020 is delivered.