FEAT: Support new data format on gov.pl
Commit brings back full functionality of the script broken with the
change of source data change.
On November 24, 2020 gov.pl site switched to a new model of data
presentation. Instead of providing a sum of cases since the beginning
of pandemic, data from November 24th was a daily delta of cases
found. In addition to that new fields were introduced.
Now JSON LocationEntity in version 1.0.0 is used to store data in old
format, and since version 1.1.0 new version of data is supported.